ai_service/requirements.txt
fastapi
uvicorn
appwrite
python-dotenv
pydantic
